Correct return type of FileHandle::size()
File size should be off_t, not size_t.

+++ platform/Stream.cpp
@@ -113,7 +113,7 @@ int Stream::sync() {
     return 0;
 }
 
-size_t Stream::size() {
+off_t Stream::size() {
     return 0;
 }
